refactor: 파일 구조 변경
로드맵 > 도메인 안에 세부적으로 파일 변경 !
