+++ src/main/java/com/simso/domain/roadmap/dto/RoadmapUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.simso.dto.roadmap;
+package com.simso.domain.roadmap.dto;
 
 import lombok.Builder;
 import lombok.Getter;
