+++ src/main/java/com/simso/domain/roadmap/entity/Roadmap.java
@@ -1,6 +1,7 @@
-package com.simso.domain;
+package com.simso.domain.roadmap.entity;
 
-import com.simso.dto.roadmap.RoadmapSaveRequestDto;
+import com.simso.domain.User;
+import com.simso.domain.roadmap.dto.RoadmapSaveRequestDto;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
