+++ src/main/java/com/simso/domain/roadmap/exception/RoadmapNotFoundException.java
@@ -1,0 +1,4 @@
+package com.simso.domain.roadmap.exception;
+
+public class RoadmapNotFoundException extends RuntimeException {
+}
